requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
asgiref==3.8.1
    # via django
django==5.1.1
    # via
    #   -r requirements.in
    #   django-ninja
django-environ==0.11.2
    # via -r requirements.in
django-ninja==1.3.0
    # via -r requirements.in
psycopg2==2.9.9
    # via -r requirements.in
pydantic==2.9.2
    # via django-ninja
pydantic-core==2.23.4
    # via pydantic
sqlparse==0.5.1
    # via django
typing-extensions==4.12.2
    # via
    #   pydantic
    #   pydantic-core
